WIP: MemCache can't work! But good point of reference
So, the story is that things have to be put in a folder! Like the role
downloaded dependencies. And trying to remove / disable all the file
based object APIs is probably a hellish nightmare. It makes more sense
to try and imitate the existing APIs but make them parallelizable (use
random UUIDs for paths and names etc.). Trying this on another branch
...